claudeagenttry2/pyproject.toml
[project]
name = "claudeagenttry2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.2",
    "claude-agent-sdk>=0.1.6",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.51.0",
]

[tool.uv.workspace]
members = [
    "claudeagenttry2",
]
